time: migrate native ABI beyond 2038
Use a signed 64-bit time_t throughout the native kernel and userland ABI while keeping the existing 32-bit UFS on-disk timestamp boundary explicit. Add TZif v2 handling, shared RTC coverage, a machine-independent SNTP ntpdate, and post-2038 runtime tests.

Fix the owning PCC preprocessor defect exposed by the ABI migration, share the required 64-bit MIPS kernel runtime across GCC and PCC, and extend the PCC regression diagnostics. Normalize non-boot administrative programs under /usr/sbin across manifests and architecture rootfs builders.

+++ include/time.h
@@ -12,7 +12,7 @@
 
 #ifndef _TIME_T
 #define _TIME_T
-typedef long time_t;
+typedef long long time_t;
 #endif
 
 #include <stddef.h>
