+++ src/libc/net/ruserpass.c
@@ -791,7 +791,7 @@ char *mkenvkey(mch)
 	while (*p)
 		*sk++ = *p++;
 	*sk++ = mch;
-	sprintf(stemp, "%ld", putmp->ut_time);
+	sprintf(stemp, "%lld", (long long)putmp->ut_time);
 	sreverse(stemp1, stemp);
 	p = stemp1;
 	while (*p)
