+++ src/libtcl/tclunxaz.c
@@ -502,7 +502,7 @@ Tcl_FileCmd(void *dummy, Tcl_Interp *interp, int argc, unsigned char **argv)
 	if (stat(fileName, &statBuf) == -1) {
 	    goto badStat;
 	}
-	sprintf(interp->result, "%ld", statBuf.st_atime);
+	sprintf(interp->result, "%lld", (long long)statBuf.st_atime);
 	return TCL_OK;
     } else if ((c == 'i') && (strncmp(argv[1], "isdirectory", length) == 0)
 	    && (length >= 3)) {
@@ -539,7 +539,7 @@ Tcl_FileCmd(void *dummy, Tcl_Interp *interp, int argc, unsigned char **argv)
 	if (stat(fileName, &statBuf) == -1) {
 	    goto badStat;
 	}
-	sprintf(interp->result, "%ld", statBuf.st_mtime);
+	sprintf(interp->result, "%lld", (long long)statBuf.st_mtime);
 	return TCL_OK;
     } else if ((c == 'o') && (strncmp(argv[1], "owned", length) == 0)) {
 	if (argc != 3) {
@@ -703,17 +703,17 @@ StoreStatData(Tcl_Interp *interp, char *varName, struct stat *statPtr)
 	    == NULL) {
 	return TCL_ERROR;
     }
-    sprintf(string, "%ld", statPtr->st_atime);
+    sprintf(string, "%lld", (long long)statPtr->st_atime);
     if (Tcl_SetVar2(interp, varName, "atime", string, TCL_LEAVE_ERR_MSG)
 	    == NULL) {
 	return TCL_ERROR;
     }
-    sprintf(string, "%ld", statPtr->st_mtime);
+    sprintf(string, "%lld", (long long)statPtr->st_mtime);
     if (Tcl_SetVar2(interp, varName, "mtime", string, TCL_LEAVE_ERR_MSG)
 	    == NULL) {
 	return TCL_ERROR;
     }
-    sprintf(string, "%ld", statPtr->st_ctime);
+    sprintf(string, "%lld", (long long)statPtr->st_ctime);
     if (Tcl_SetVar2(interp, varName, "ctime", string, TCL_LEAVE_ERR_MSG)
 	    == NULL) {
 	return TCL_ERROR;
